HF_SPACE/requirements.txt
diffusers==0.21.4
datasets
wandb
accelerate
gradio
huggingface_hub
torch 
numpy 
torchvision 
pillow